requirements.txt
# Core dependencies
crawl4ai>=0.3.74
playwright>=1.40.0
browser-use>=0.1.0
openai>=1.12.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=13.0.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0